refactor: move loadCommands into index and remove unused code
