+++ package.json
@@ -42,7 +42,6 @@
   },
   "dependencies": {
     "abort-controller": "^3.0.0",
-    "async": "^2.6.1",
     "async-iterator-all": "^1.0.0",
     "async-iterator-to-pull-stream": "^1.3.0",
     "bignumber.js": "^9.0.0",
@@ -51,28 +50,20 @@
     "buffer": "^5.4.2",
     "callbackify": "^1.1.0",
     "cids": "~0.7.1",
-    "concat-stream": "github:hugomrdias/concat-stream#feat/smaller",
     "debug": "^4.1.0",
-    "detect-node": "^2.0.4",
     "err-code": "^2.0.0",
     "explain-error": "^1.0.4",
-    "flatmap": "0.0.3",
     "form-data": "^3.0.0",
-    "glob": "^7.1.3",
     "ipfs-block": "~0.8.1",
     "ipfs-utils": "^0.4.0",
     "ipld-dag-cbor": "~0.15.0",
     "ipld-dag-pb": "^0.18.1",
     "ipld-raw": "^4.0.0",
     "is-ipfs": "~0.6.1",
-    "is-pull-stream": "0.0.0",
-    "is-stream": "^2.0.0",
-    "iso-stream-http": "~0.1.2",
     "it-glob": "0.0.6",
     "it-tar": "^1.1.0",
     "it-to-stream": "^0.1.1",
     "iterable-ndjson": "^1.1.0",
-    "kind-of": "^6.0.2",
     "ky": "^0.15.0",
     "ky-universal": "^0.3.0",
     "merge-options": "^2.0.0",
@@ -80,20 +71,10 @@
     "multibase": "~0.6.0",
     "multicodec": "~0.5.1",
     "multihashes": "~0.4.14",
-    "ndjson": "github:hugomrdias/ndjson#feat/readable-stream3",
-    "once": "^1.4.0",
     "parse-duration": "^0.1.1",
     "peer-id": "~0.12.3",
     "peer-info": "~0.15.1",
-    "promise-nodeify": "^3.0.1",
-    "promisify-es6": "^1.0.3",
-    "pull-defer": "~0.2.3",
-    "pull-stream": "^3.6.9",
-    "pull-to-stream": "~0.1.1",
-    "pump": "^3.0.0",
-    "qs": "^6.5.2",
-    "readable-stream": "^3.1.1",
-    "stream-to-pull-stream": "^1.7.2"
+    "promise-nodeify": "^3.0.1"
   },
   "devDependencies": {
     "aegir": "^20.4.1",
@@ -103,6 +84,8 @@
     "interface-ipfs-core": "^0.121.0",
     "ipfsd-ctl": "^0.47.1",
     "nock": "^11.4.0",
+    "promisify-es6": "^1.0.3",
+    "pull-stream": "^3.6.14",
     "stream-equal": "^1.1.1"
   },
   "engines": {
