+++ src/ls.js
@@ -1,23 +1,15 @@
 'use strict'
 
-const IsIpfs = require('is-ipfs')
-const cleanCID = require('./utils/clean-cid')
+const { Buffer } = require('buffer')
+const CID = require('cids')
 const configure = require('./lib/configure')
 
 module.exports = configure(({ ky }) => {
   return async function * ls (path, options) {
     options = options || {}
 
-    try {
-      path = cleanCID(path)
-    } catch (err) {
-      if (!IsIpfs.ipfsPath(path)) {
-        throw err
-      }
-    }
-
     const searchParams = new URLSearchParams()
-    searchParams.set('arg', path.toString())
+    searchParams.set('arg', `${Buffer.isBuffer(path) ? new CID(path) : path}`)
 
     if (options.long !== undefined) {
       searchParams.set('long', options.long)
